requirements.txt
fastapi
vllm
pydantic
uvicorn
pathlib
pyyaml
flashinfer-python==0.2.4
watchdog